+++ WFInfo/Components/NumericOnlyEntry.cs
@@ -3,7 +3,7 @@ using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Input;
 
-namespace WFInfo
+namespace WFInfo.Components
 {
     public static class NumericOnlyEntry
     {
