+++ WFInfo/Data.cs
@@ -13,6 +13,7 @@ using System.Security.Cryptography;
 using System.Threading;
 using System.Threading.Tasks;
 using WebSocketSharp;
+using WFInfo.Settings;
 
 namespace WFInfo
 {
