+++ WFInfo/Overlay.xaml.cs
@@ -5,6 +5,7 @@ using System.Windows;
 using System.Windows.Interop;
 using System.Windows.Media;
 using System.Windows.Threading;
+using WFInfo.Settings;
 
 namespace WFInfo
 {
